refactor(admin): move control routes
